Add test-unit as test group dependency
When we run 'rake test' through bundler like 'bundle exec rake test', we
need test-unit gem installed.  It's because testrb in bundler gem
activates 'test-unit' gem in contrast stock testrb in Ruby dist doesn't.

I'm suspecting it's a bug of bundler but it might be good to add
'test-unit' to Gemfile for a workaround because Gemfile is for bundler.